all code optimized

+++ Project/Project/Controllers/BusesController.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Forms;
 
 namespace Project.Controllers
 {
@@ -16,6 +17,11 @@ namespace Project.Controllers
         }
         public static bool addBus(dynamic bus)
         {
+            if (bus.coach.Length == 0 || bus.type.Length == 0)
+            {
+                MessageBox.Show("Fill all the required fields");
+                return false;
+            }
             return db.Buses.addBus(bus);
         }
         public static Bus searchBus(string coach)
@@ -24,10 +30,25 @@ namespace Project.Controllers
         }
         public static bool updateBus(dynamic bus)
         {
+            if (bus.id == 0)
+            {
+                MessageBox.Show("Search a bus first");
+                return false;
+            }
+            if (bus.coach.Length == 0 || bus.type.Length == 0)
+            {
+                MessageBox.Show("Fill all the required fields");
+                return false;
+            }
             return db.Buses.updateBus(bus);
         }
         public static bool deleteBus(int id)
         {
+            if (id == 0)
+            {
+                MessageBox.Show("Search a bus first");
+                return false;
+            }
             return db.Buses.deleteBus(id);
         }
     }
