+++ Project/Project/Controllers/SalesmanController.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Forms;
 
 namespace Project.Controllers
 {
@@ -17,6 +18,17 @@ namespace Project.Controllers
         }
         public static bool addSalesman(dynamic salesman)
         {
+            if (salesman.name.Length == 0 || salesman.username.Length == 0 || salesman.password.Length == 0)
+            {
+                MessageBox.Show("Fill all the required field");
+                return false;
+            }
+            var has = searchSalesman(salesman.username);
+            if (has != null)
+            {
+                MessageBox.Show("Username already used");
+                return false;
+            }
             return db.Sellers.addSalesMan(salesman);
         }
         public static Salesman searchSalesman(string username)
@@ -25,10 +37,28 @@ namespace Project.Controllers
         }
         public static bool updateSalesman(dynamic salesman)
         {
+            if (salesman.id == 0)
+            {
+                MessageBox.Show("Search a Salesman First");
+                return false;
+            }
+            if (salesman.name.Length == 0 || salesman.username.Length == 0 || salesman.password.Length == 0)
+            {
+                MessageBox.Show("Fill all the required fields");
+                return false;
+            }
+            var has = searchSalesman(salesman.username);
+            if (has != null) { MessageBox.Show("Username already exists"); return false; }
             return db.Sellers.updateSalesman(salesman);
         }
         public static bool deleteSalesman(int id)
         {
+            if (id == 0)
+            {
+                MessageBox.Show("Search a Salesman First");
+                return false;
+            }
+
             return db.Sellers.deleteSalesman(id);
         }
     }
